Added frontend select exam and level modal
