+++ src/api/utils/types.ts
@@ -10,7 +10,7 @@ declare global {
 export type Subject = {
   subject_id: string;
   subject_name: string;
-  subject_curriculum: string;
+  subject_board: string;
   subject_level: string;
   subject_icon: string;
 };
