+++ src/utils/subjects.ts
@@ -3,126 +3,126 @@ export const subjects = [
     'subject_id': '0452',
     'subject_name': 'Accounting',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '💰',
   },
   {
     'subject_id': '0400',
     'subject_name': 'Art & Design',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '🎨',
   },
   {
     'subject_id': '0610',
     'subject_name': 'Biology',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '🔬',
   },
   {
     'subject_id': '0450',
     'subject_name': 'Business Studies',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '💼',
   },
   {
     'subject_id': '0620',
     'subject_name': 'Chemistry',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '🧪',
   },
   {
     'subject_id': '0478',
     'subject_name': 'Computer Science',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '💻',
   },
   {
     'subject_id': '0445',
     'subject_name': 'Design & Technology',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '🛠️',
   },
   {
     'subject_id': '0411',
     'subject_name': 'Drama',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '🎭',
   },
   {
     'subject_id': '0455',
     'subject_name': 'Economics',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '💰',
   },
   {
     'subject_id': '0475',
     'subject_name': 'English - Literature',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '📚',
   },
   {
     'subject_id': '0500',
     'subject_name': 'English - Language',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '📚',
   },
   {
     'subject_id': '0460',
     'subject_name': 'Geography',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '🌎',
   },
   {
     'subject_id': '0470',
     'subject_name': 'History',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '📜',
   },
   {
     'subject_id': '0580',
     'subject_name': 'Mathematics',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '📉',
   },
   {
     'subject_id': '0980',
     'subject_name': 'Core Mathematics',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '📊',
   },
   {
     'subject_id': '0410',
     'subject_name': 'Music',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '🎵',
   },
   {
     'subject_id': '0413',
     'subject_name': 'Physical Education',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '⚽️',
   },
   {
     'subject_id': '0625',
     'subject_name': 'Physics',
     'subject_level': 'IGCSE',
-    'subject_curriculum': 'Cambridge',
+    'subject_board': 'Cambridge',
     'subject_icon': '🚀',
   },
 ];
